Wrapper classes do not and will not have "Wrapper" in their signature.
